c1541: fix warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38382 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c1541.c
@@ -5032,7 +5032,7 @@ static int write_cmd(int nargs, char **args)
         unsigned int length = fileio_get_bytes_left(finfo);
         /* Records and positions start counting at 1 */
         unsigned int max_record =
-            ((length + rel_record_length - 1) / rel_record_length);
+            (unsigned int)((length + rel_record_length - 1) / rel_record_length);
         unsigned int record;
         int err;
